day 7 part 2
I'm actually really proud of this: part 2 took a while to write, but
once it finally came together, it got the right answer on the first
run. That was fun!

There's a big TODO to split out the outputs and halts properly, but
I'm not sure it will ever actually get done. Refactoring previous
days' exercise parts is a pretty low priority, unfortunately.

It may have been possible to solve this all single-threaded using
async-await, which is designed for _precisely_ this use case. I
just don't know about how that all works to feel comfortable
attempting that yet. It would be a good TODO.

The funny thing today is that I remember, in Day 5 when they
introduced the inputs and outputs system, thinking through how
to architect it. Just using stdin/stdout seemed like the obvious
solution, but I had a feeling that if nothing else, it would kill
my benchmarks. "Channels are probably the right choice here," I
thought to myself, "but that seems like a lot of work. I may as
well just use vectors."

It wasn't a huge petard to hoist myself up, but I still did it.