day 8 part 2
this was weirdly easy, I think. I suppose I could have upped the ante
by OCRing the part 2 output, but that doesn't seem that productive
a use of my time.

Hypothesis: AoC attempts to have a slow, steady increase in difficulty
during the week, with spikes high on saturday and spikes low on sunday.