bugfix: account for velocity
What's really interesting is that this fix works perfectly if we
account for velocity, _and_ if we start the step count from 1.

Starting the step count from 1 makes perfect sense, that's how
it worked in my first implementation.

I wonder if it's a necessary property that you get the right answer
either by starting from 2, or by ignoring the velocity entirely?
If no, it's weird that it happened to work perfectly for both the
real input and both test inputs. If yes, that seems like a useful,
though abstruse, optimization.