bugfix: get the robot navigating such that it finds the target
Unfortunately, it doesn't generate the correct path length. I
can't say with certainty what's going on yet, but my suspicion:
the direct path it's finding goes through unknown tiles, some
of which are walls.

Might be worth adding a cartography display module, but might
be faster to have the robot path back and forth between the origin
and the target until the path distance is consistent. This should
be safe, because in the event that it takes a roundabout route,
the return trip should attempt new unexplored paths over the
known roundabout way.

Note: resetting by teleportation is not safe, because we can't
teleport in the intcode machine.