add incomplete algorithm to find actual shortest path
Something's not working right with this implementation, and I don't
yet know what. It's spinning permanently when trying to do the shuttle
back and forth between the target point and the origin.
It's probably time to implement a debug helper which
can print out the internal map, suitably trimmed; if nothing else, that
would help me validate the path calculations.