add cartography, partially debug apparent freeze
It turns out that it takes a lot of attempts, and a lot of re-
planning, to get back to the origin: over 512 replanning attempts
are required for my personal input. A* is a reasonably fast maze-
solver, but almost all of these failed attempts are used to discover
all the relevant walls of the maze. Release mode helps a lot here.

At this point, not everything works properly yet: in the function
find_shortest_path_to_origin, we assume that once we've found the
shortest path in one direction, reversing the route will Just Work.
The panics the program is currently launching indicate that this
is not a safe assumption.

Next up: require a successful, no-replanning traversal twice in a
row before assuming that we know enough of the map to plan a traversal.