finish day 15 part 2
Now I feel really smart with the minute-tracking system I came up with.
It worked fine! I really was off by 1.

Filling in the map first makes both parts fast:

$ ./bench.sh inputs/15.txt --day 15
Sun 31 May 2020 06:50:28 PM CEST
+ hyperfine 'target/release/aoc2019 inputs/15.txt --day 15' 'target/release/aoc2019 inputs/15.txt --no-part1 --part2 --day 15'
Benchmark #1: target/release/aoc2019 inputs/15.txt --day 15
  Time (mean ± σ):      10.1 ms ±   0.4 ms    [User: 17.8 ms, System: 0.6 ms]
  Range (min … max):     9.6 ms …  14.6 ms    200 runs

  Warning: The first benchmarking run for this command was significantly slower than the rest (14.6 ms). This could be caused by (filesystem) caches that were not filled until after the first run. You should consider using the '--warmup' option to fill those caches before the actual benchmark. Alternatively, use the '--prepare' option to clear the caches before each timing run.

Benchmark #2: target/release/aoc2019 inputs/15.txt --no-part1 --part2 --day 15
  Time (mean ± σ):       9.3 ms ±   0.3 ms    [User: 16.7 ms, System: 0.6 ms]
  Range (min … max):     8.8 ms …  10.2 ms    316 runs

Summary
  'target/release/aoc2019 inputs/15.txt --no-part1 --part2 --day 15' ran
    1.08 ± 0.06 times faster than 'target/release/aoc2019 inputs/15.txt --day 15'

Part 2 is of course faster than part 1 because in part 2, the robot doesn't
need to move anymore; it just sits in place and thinks about the layout
of the map. That frees it of the need to interact with the intcode side
of things at all. That it's only a small bit faster than part 1 suggests
that quite a large portion of the overall time is in discovering the oxygenator
and filling in the map in the first place.