big refactor: extract map struct into immutable datatype
Tests pass again, and we now have a generic Map type which
should be generally useful for 2d stuff in the future.

Unfortunately, it looks like this isn't actually where the bulk of
the runtime was being spent; it still takes more time than I want
to bother with for part 1 to run. Given that the new map type
is cheap to clone, maybe the issue is all the mutation?

Obviously, guessing isn't all that helpful here; real profiling
would be the answer. That's hard, though, so guessing it is!